Added two experimental functions:
TrackNTraceBatch
we could use a function that can be called without invoking the gui, for
example when all settings are known and a large batch of movies has to be
handled overnight. TrackNTraceBatch needs a cell array of full movie
paths, dark_img, all the options structs and then just loops over
everything. it's not commented yet.

filterPositionsSNB
nothing gets filtered right now. there should be a possible filtering step
before tracking at least. filterPositionsSNB changes the exit flag of
particles with A/B < SNB to -3 so they get removed right before starting
tracking.
filtering is currently tested in TrackNTraceBatch which accepts a struct
filterOptions containing a function string fun and a cell array of
arguments args. later, the fun is called with all the arguments using
eval()