Major update to insert global variables
globalOptions, imgCorrect and movie are now global, allowing plugin
functions access to them whenever needed. All plugins now will know the
current movie frame and can, if necessary, get adjacent frames and use
those for localization in an arbitrary way.
For this purpose, a correctMovie function was added which can be used to
correct frames from "movie" with imgCorrect and/or photon conversion.