Fixed "missing functionHandle" design bug + small fixes
(x) Due to the way "createOptionsPanel"  was implemented, the "options"
used by the callbacks of the plugin were not synced correctly with the
appdata options (getappdata(h_panel,'options'). Thus, any change to
options outside the callbacks was overwritten most of the time. As the
functionHandle was defined outside "createOptionsPanel" it was often
missing.
(x) Fixed settings of the currently selected plugins not beeing
remembered if settings are loaded.
(x) Cleaned a lot of redundant updates of plugins in the settings GUI.
  This reduces flickering and improves how the GUI feeles, especially
  on slow machines.
(x) Added comments to the settings GUI