Introducing description of output parameters for plugins
- Made all options global so that they can be accessed by every plugin
  This is needed so that e.g. a tracking plugin can read the name of the
output parameters from the fitting plugin, if it simply wants to "pass
on" parameters to its output
- runPreview uses options via global variables now, as otherwise changes
in one plugins options are not visible via the global interface.
Although this is somewhat bad coding, it can hardly be avoided
without making TNT into a class/object.
- Visualizer for fittingData can now show histograms for data in all
columns of fittingData using the appropriate names as specified by the
fitting plugin.

Other:
- Few cosmetical changes / cleanup
- Disabled backtrace on few warnings to make output less cluttered
where appropriate.
- Removed superfluous nn_tracker_cpp.m file
- Renamed testTrackerSettings.m to runPreview.m
- Renamed fitData to fittingData, so that we always have xOptions/xData
pairs with same name 'x'.
- Settings GUI only displays warnings while loading plugins the first
time (i.e. when not in 'single file mode').