Changed behaviour of globalOptions.lastFrame
Before, if "inf" was set in the GUI, this was corrected to be the last
frame of the movie that actually exists during processing.

Due to this, loading settings from an existing movie set lastFrame to
the last frame of that movie, even if the user did not change the
setting from "inf" for the other movie. This causes potential confusion.

To preserve the information about the last frames number, this is now
saved to the variable firstFrame_lastFrame.