Fixed unexpected behaviour of TNT NearestNeighbor tracker.
The minTrajLength parameter was ignored if gap closing was disabled (by
having 0 frames gap closing). Without gap closing minTrajLength and
minSegLength perform essentially the same task, so this is not a big
deal. But as it might confuse the user, filtering is now performed
before and after gap closing, even if gap closing is disabled.