Fixed error on startup based on plugin names
If two plugins exist where one plugins name is a subtring of the other
and the subtring one is set as the default plugin, e.g. "Cross correlation" & "Cross correlation Airy" this caused trouble
in the TNTgui when trying to select the default plugin and it crashed.
This is fixed.

Also fixed some legacy code where "fittingData" was apparently not
renamed to its current name "refinementData"